Refactor moving PresentationView to views/presentation
